Enable NarrowingCompoundAssignment error-prone check
In Java, compound assignment with wider type on the right-hand-side is a
silent truncation.

This enables error-prone check for such cases and refactors code to make
it compliant. Some places are updated based solely on the assumption
that they weren't incorrect.

+++ lib/trino-filesystem/src/test/java/io/trino/filesystem/AbstractTestTrinoFileSystem.java
@@ -274,7 +274,7 @@ public abstract class AbstractTestTrinoFileSystem
 
                 // skip 1MB at a time
                 inputStream.seek(0);
-                int expectedPosition = 0;
+                long expectedPosition = 0;
                 for (int i = 0; i < 15; i++) {
                     long skipSize = inputStream.skip(MEGABYTE);
                     assertThat(skipSize).isEqualTo(MEGABYTE);
